requirements.txt
asgiref==3.2.10
backcall==0.2.0
colorama==0.4.4
decorator==4.4.2
Django==3.1.2
django-extensions==3.0.9
djangorestframework==3.12.1
ipython==7.18.1
ipython-genutils==0.2.0
jedi==0.17.2
mysqlclient==2.0.1
parso==0.7.1
pickleshare==0.7.5
prompt-toolkit==3.0.8
Pygments==2.7.2
pytz==2020.1
sqlparse==0.4.1
traitlets==5.0.5
wcwidth==0.2.5
django-cors-headers==3.5.0
djangorestframework-jwt==1.11.0
django-filter==2.4.0 